refactor(runner): extract session history telemetry
